+++ tests/nativeapp/test_artifacts.py
@@ -1,7 +1,9 @@
+import os
 from pathlib import Path
 from typing import List, Optional
 
 import pytest
+from click.exceptions import ClickException
 from snowflake.cli.api.project.definition import load_project_definition
 from snowflake.cli.plugins.nativeapp.artifacts import (
     ArtifactMapping,
@@ -11,8 +13,12 @@ from snowflake.cli.plugins.nativeapp.artifacts import (
     SourceNotFoundError,
     TooManyFilesError,
     build_bundle,
+    resolve_without_follow,
+    source_path_to_deploy_path,
     translate_artifact,
 )
+
+from tests.nativeapp.utils import touch
 
 
 def trimmed_contents(path: Path) -> Optional[str]:
@@ -163,3 +169,60 @@ def test_too_many_files(project_definition_files):
                 ArtifactMapping("app/streamlit/*.py", "somehow_combined_streamlits.py")
             ],
         )
+
+
+@pytest.mark.parametrize(
+    "project_path,expected_path",
+    [
+        [
+            "srcfile",
+            "deploy/file",
+        ],
+        [
+            "srcdir",
+            "deploy/dir",
+        ],
+        [
+            "srcdir/nested_file1",
+            "deploy/dir/nested_file1",
+        ],
+        [
+            "srcdir/nested_dir/nested_file2",
+            "deploy/dir/nested_dir/nested_file2",
+        ],
+        [
+            "srcdir/nested_dir",
+            "deploy/dir/nested_dir",
+        ],
+        [
+            "not-in-deploy",
+            None,
+        ],
+    ],
+)
+def test_source_path_to_deploy_path(
+    temp_dir,
+    project_path,
+    expected_path,
+):
+    # Source files
+    touch("srcfile")
+    touch("srcdir/nested_file1")
+    touch("srcdir/nested_dir/nested_file2")
+    touch("not-in-deploy")
+    # Build
+    os.mkdir("deploy")
+    os.symlink("srcfile", "deploy/file")
+    os.symlink(Path("srcdir").resolve(), Path("deploy/dir"))
+
+    files_mapping = {
+        Path("srcdir").resolve(): resolve_without_follow(Path("deploy/dir")),
+        Path("srcfile").resolve(): resolve_without_follow(Path("deploy/file")),
+    }
+
+    if expected_path:
+        result = source_path_to_deploy_path(Path(project_path).resolve(), files_mapping)
+        assert result == resolve_without_follow(Path(expected_path))
+    else:
+        with pytest.raises(ClickException):
+            source_path_to_deploy_path(Path(project_path).resolve(), files_mapping)