+++ RELEASE-NOTES.md
@@ -6,9 +6,14 @@
 
 ## New additions
 * `snow sql` command supports now client-side templating of queries.
+* New `snow app deploy` functionality:
+  * Passing files and directories as arguments syncs these only: `snow app deploy some-file some-dir`.
+  * `--recursive` syncs all files and subdirectories recursively.
+  * `--prune` deletes specified files from the stage if they don't exist locally.
 
 ## Fixes and improvements
 * More human-friendly errors in case of corrupted `config.toml` file.
+* Fixed a bug in `snow app` that caused files to be re-uploaded unnecessarily.
 
 # v2.2.0